add extend view
